requirements.txt
pygame==2.0.0
Pillow==2.2.2